pyproject.toml
[tool.poetry]
name = "optunabatch"
version = "0.1.0"
description = ""
authors = ["richwomanbtc <richwomanbtc@gmail.com>"]
readme = "README.md"
packages = [{ include = "optunabatch" }]

[tool.poetry.dependencies]
python = "^3.11"
optuna = "^3.6.1"
google-cloud-batch = "^0.17.18"
google-cloud-pubsub = "^2.21.1"
google-cloud-storage = "^2.16.0"
pyhumps = "^3.8.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
